day22: impl part2
Initially tried to be smart and wasted a day on it.
Took a break, then tried to solve it by splitting blocks in half when they
had collissions. Failed. It either creates too many chunks or the code is broken.

Finally implemented a solution to split on the intersection planes. This
did the trick, finally!!! This was the thing I tried to be smart about and
find a short cut to it.... Will I ever learn that lesson... Keep it simple!!!!